Merged PR 1053: Tenancy Resolver
Added a tenancy resolver that allows querying and inserting entities that require tenancy id. It will automatically set the tenant id from the request header.

So no need to specify school id on every endpoint created. once a request is made from any school admin, staff or student, the tenant id in the header is used to query or create the entity table.